added emojipicker, fixed deletion of saving goals

+++ src/components/savinggoal/SavingGoalCard.vue
@@ -9,9 +9,12 @@
     </div>
     <div v-else>
       <h2>{{ savingGoal.name }}</h2>
-      <p><strong>Sparemål:</strong> {{ savingGoal.targetAmount }} kr</p>
+      <p><strong>Sparemål:</strong> {{ savingGoal.savedAmount }} / {{ savingGoal.targetAmount }} kr</p>
       <p><strong>Frist:</strong> {{ savingGoal.deadline }}</p>
-      <img v-if="savingGoal.mediaUrl" :src="savingGoal.mediaUrl" alt="Media" class="img">
+      <img v-if="savingGoal.mediaUrl?.length && savingGoal.mediaUrl.length > 4" :src="savingGoal.mediaUrl" alt="Media" class="img">
+      <div v-else>
+        <span class="emoji">{{ savingGoal.mediaUrl }}</span>
+      </div>
       <p v-if="savingGoal.completed" class="completed">Completed</p>
       <div class="delete-icon" @click="confirmDelete">✖</div>
       <div class="edit-icon" @click="startEditing">✎</div>
@@ -22,12 +25,13 @@
 <script setup lang="ts">
 import { defineProps, defineEmits, ref } from 'vue';
 import type { SavingGoal } from '@/types/SavingGoal';
-import { updateSavingGoal, deleteSavingGoal } from '@/api/savingGoalHooks';
+import { updateSavingGoal, deleteSavingGoal, deleteSavingGoalFromUser } from '@/api/savingGoalHooks';
+import { useUserStore } from '@/stores/userStore';
 
 const props = defineProps({
   savingGoal: {
     type: Object as () => SavingGoal,
-    default: () => ({ name: '', targetAmount: 0, deadline: '', mediaUrl: '', completed: false })
+    default: () => ({ name: '', targetAmount: 0, savedAmount: '0', deadline: '', mediaUrl: '', completed: false })
   }
 });
 
@@ -35,6 +39,8 @@ const emits = defineEmits(['deleteGoal', 'updateGoal']);
 
 const isEditing = ref(false);
 const editableGoal = ref({ ...props.savingGoal }); 
+const userStore = useUserStore();
+const userId = userStore.getUserId;
 
 const startEditing = () => {
   isEditing.value = true;
@@ -53,11 +59,19 @@ const saveChanges = async () => {
 
 const confirmDelete = async () => {
   try {
-    await deleteSavingGoal(Number(props.savingGoal.id));
+    const confirmDelete = confirm('Are you sure you want to delete this saving goal?');
+    if (!confirmDelete) return;
+    await deleteSavingGoalFromUser(userId, Number(props.savingGoal.id));
+    try {
+      await deleteSavingGoal(Number(props.savingGoal.id));
+    } catch (error) {
+      console.log("Can't delete saving goal that has users");
+    }
     emits('deleteGoal', props.savingGoal.id);
   } catch (error) {
     console.error('Error deleting saving goal:', error);
   }
+  window.location.reload();
 };
 </script>
   
@@ -65,6 +79,8 @@ const confirmDelete = async () => {
 .saving-goal-card {
   position: relative;
   max-width: 600px;
+  width: 222px;
+  height: 365px;
   margin: auto;
   margin-top: 2%;
   padding: 20px;
@@ -89,7 +105,12 @@ const confirmDelete = async () => {
 
 .img {
   max-width: 30vh;
+  width: 30vh;
   margin-top: 10px;
+}
+
+.emoji {
+  font-size: 22vh;
 }
 
 .completed {
